move composable previews into another file in order to exclude it from jacoco coverage
